refactor: Extract PlayHistoryTracker from AudioManager
AudioManager had 1302 lines with play history tracking mixed into source
orchestration. Extract all play history logic into a dedicated
PlayHistoryTracker class (493 lines), reducing AudioManager to 882 lines.

Extracted: RecordPlayStartAsync, OnTrackIdentified, OnSongChanged,
OnBluetoothMetadataChanged, GetSourceMetadata, and all helper methods.
Uses Func<IAudioSource?> delegate to avoid circular DI dependency.

All 1313 tests pass, verified play history recording on Ubuntu deployment.
